update to module structure
