+++ mom6/stats/mom6_regrid_tercile.py
@@ -4,11 +4,14 @@ REGRIDDED forecast/hindcast data.
 
 """
 # %%
+import os
 import sys
 import warnings
 import xarray as xr
 from dask.distributed import Client
-from mom6_regrid import mom6_hindcast,mom6_encoding_attr
+from mom6 import DATA_PATH
+from mom6.mom6_module import mom6_process as mp
+# from mom6_regrid import mom6_hindcast,mom6_encoding_attr
 warnings.simplefilter("ignore")
 
 # %%
@@ -25,10 +28,10 @@ if __name__=="__main__":
     else:
         varname = sys.argv[1]
 
-    # data locations    
-    mom6_dir = "/Datasets.private/regional_mom6/hindcast/regrid/"
-    mom6_tercile_dir = "/Datasets.private/regional_mom6/tercile_calculation/regrid/"
-    file_list = mom6_hindcast(mom6_dir)
+    # data locations
+    mom6_dir = os.path.join(DATA_PATH,"hindcast/regrid/")
+    mom6_tercile_dir = os.path.join(DATA_PATH,"tercile_calculation/regrid/")
+    file_list = mp.MOM6Misc.mom6_hindcast(mom6_dir)
     var_file_list = []
     for file in file_list :
         if varname in file :
@@ -48,7 +51,7 @@ if __name__=="__main__":
                 keep_attrs=True
             )
         )
-     
+
         ds_tercile = xr.Dataset()
         ds_tercile.attrs = ds.attrs
         ds_tercile['f_lowmid'] = da_tercile.isel(quantile=0)
@@ -57,7 +60,7 @@ if __name__=="__main__":
 
         # output the netcdf file
         print(f'output {mom6_tercile_dir}{file[len(mom6_dir):]}')
-        mom6_encoding_attr(
+        mp.MOM6Misc.mom6_encoding_attr(
                 ds,
                 ds_tercile,
                 var_names=list(ds_tercile.keys()),
