+++ mom6/stats/mom6_tercile_probability.py
@@ -5,13 +5,7 @@ This is the script to genearte the tercile plot based on the
 forecast(hindcast) that is generated by Andrew Ross at GFDL.
 
 """
-from typing import (
-    Literal,
-    List
-)
-from datetime import date
-from dateutil.relativedelta import relativedelta
-import cftime
+import os
 import numpy as np
 import pandas as pd
 import xarray as xr
@@ -20,220 +14,8 @@ from matplotlib import cm
 import matplotlib.pyplot as plt
 import cartopy.crs as ccrs
 from scipy.stats import norm as normal
-from mom6_regrid import mom6_hindcast
+from mom6.mom6_module import mom6_process as mp
 
-
-def get_mom6_regionl_mask() -> xr.Dataset:
-    """return the EPU mask in the mom6 grid
-    """
-    ds = xr.open_dataset('/Datasets.private/regional_mom6/masks/region_masks.nc')
-    return ds.set_coords(['geolon','geolat'])
-
-def get_mom6_mask(
-    mask : Literal['wet','wet_c','wet_u','wet_v'] = 'wet',
-    source : Literal['raw','regrid'] = 'raw'
-) -> xr.DataArray:
-    """
-    The function is designed to export the various mask provided
-    on the MOM6 grid from the ocean_static.nc file
-
-    Parameters
-    ----------
-    mask : str
-        The mask name based on the variable name in the ocean_static.nc file.
-        It has the following options 1. wet (0 if land, 1 if ocean at
-        tracer points), 2. wet_c (0 if land, 1 if ocean at corner (Bu)
-        points), 3. wet_u (0 if land, 1 if ocean at zonal velocity (Cu) 
-        points), 4. wet_v (0 if land, 1 if ocean at meridional velocity
-        (Cv) points), by default 'wet'.
-    
-    source : Literal[&#39;raw&#39;,&#39;regrid&#39;], optional
-        The data extracted should be the regridded result or 
-        the original model grid (curvilinear), by default 'raw'
-
-    Returns
-    -------
-    xr.DataArray
-        The Xarray DataArray object that represent the ocean mask.
-    """
-    if source == 'raw':
-        ds = xr.open_dataset('/Datasets.private/regional_mom6/static/ocean_static.nc')
-        da = ds.set_coords(['geolon','geolat'])[mask]
-    elif source == 'regrid':
-        ds = xr.open_dataset('/Datasets.private/regional_mom6/static/regrid/ocean_static.wet.nc')
-        da = ds[mask]
-    return da
-
-
-def get_mom6(
-    iyear : int,
-    imonth : int,
-    var : str,
-    source : Literal['raw','regrid'] = 'raw'
-) -> xr.Dataset:
-    """
-    Return the mom6 rawgrid/regridded hindcast/forecast field
-    with the static field combined and setting the
-    lon lat related variables to coordinate 
-
-    Parameters
-    ----------
-    imonth : int
-        initial year of forecast
-    imonth : int
-        initial month
-    var : str
-        variable name one want to exetract from the data
-    source : Literal[&#39;raw&#39;,&#39;regrid&#39;], optional
-        The data extracted should be the regridded result or 
-        the original model grid (curvilinear), by default 'raw'
-
-    Returns
-    -------
-    xr.Dataset
-        The Xarray Dataset object is the merged dataset of
-        all forecast field include in the `file_list`. The
-        Dataset object is lazily-loaded.
-    """
-    if source == 'raw' :
-        # getting the forecast/hindcast data
-        mom6_dir = "/Datasets.private/regional_mom6/hindcast/"
-        file_list = mom6_hindcast(mom6_dir)
-
-        # static field
-        ds_static = xr.open_dataset('/Datasets.private/regional_mom6/ocean_static.nc')
-        ds_static = ds_static.set_coords(
-            ['geolon','geolat',
-            'geolon_c','geolat_c',
-            'geolon_u','geolat_u',
-            'geolon_v','geolat_v']
-        )
-
-        # merge the static field with the variables
-        for file in file_list:
-            #iyear_flag = f'i{iyear}' in file
-            imon_flag = f'i{imonth}' in file
-            var_flag = var in file
-            if imon_flag and var_flag :
-                ds = xr.open_dataset(file).sel(init=f'{iyear}-{imonth}')
-
-        ds = xr.merge([ds_static,ds])
-
-    elif source == 'regrid':
-        # getting the forecast/hindcast data
-        mom6_dir = "/Datasets.private/regional_mom6/hindcast/regrid/"
-        file_list = mom6_hindcast(mom6_dir)
-
-        # read only the needed file
-        for file in file_list:
-            #iyear_flag = f'i{iyear}' in file
-            imon_flag = f'i{imonth}' in file
-            var_flag = var in file
-            if imon_flag and var_flag :
-                ds = xr.open_dataset(file).sel(init=f'{iyear}-{imonth}')
-
-    return ds
-
-def get_mom6_tercile(
-        imonth : int,
-        var : str,
-        source : Literal['raw','regrid'] = 'raw'
-) -> xr.Dataset:
-    """return the mom6 quantile from the forecast
-
-    Parameters
-    ----------
-    imonth : int
-        The initialization month that generate the forecast
-    var : str
-        variable name one want to exetract from the data
-    source : Literal[&#39;raw&#39;,&#39;regrid&#39;], optional
-        The data extracted should be the regridded result or 
-        the original model grid (curvilinear), by default 'raw'
-
-    Returns
-    -------
-    xr.Dataset
-        A dataset that include the f_lowmid and f_midhigh value which 
-        represent SST values at the boundaries between the terciles. 
-        `f_lowmid` represent the boundary value between lower and middle
-        tercile. `f_midhigh` represent the boundary value between middle
-        and upper tercile. (the filename 'quantile' MIGHT be error naming)
-    """
-    if source == 'raw':
-        # getting the forecast/hindcast tercile data
-        mom6_dir = "/Datasets.private/regional_mom6/tercile_calculation/"
-        return xr.open_dataset(f'{mom6_dir}/forecast_quantiles_i{imonth:02d}.nc')
-
-    elif source == 'regrid':
-        # getting the regridd forecast/hindcast tercile data
-        mom6_dir = "/Datasets.private/regional_mom6/tercile_calculation/regrid/"
-        return xr.open_dataset(f'{mom6_dir}/{var}_forecasts_i{imonth}.nc')
-
-
-def get_init_fcst_time(
-    iyear : int,
-    imonth : int,
-    bins : List[int],
-) -> dict:
-    """_summary_
-
-    Parameters
-    ----------
-    iyear : int
-        The year of initialization time
-    imonth : int
-        The month of initialization time
-    bins : List[int]
-        The `lead_bin` used to binned the leading month result
-        example is `lead_bins = [0, 3, 6, 9, 12]` for four seasonal
-        mean.
-
-    Returns
-    -------
-    dict
-        with two key-value pairs, 'init': initial_time and
-        'fcst': mean forecasted during the binned period 
-    """
-    # get the cftime of initial time
-    btime = cftime.datetime(iyear,imonth,1)
-
-    # store the forecast time format based on all leadtime
-    forecasttime = []
-    period_length = bins[1]-bins[0]-1  # assuming the bins are equal space
-    for l in range(0,12):
-        # leadtime period start
-        sdate = (
-            date.fromisoformat(f'{btime.year}-'+
-                                f'{btime.month:02d}-'+
-                                f'{1:02d}')
-            +relativedelta(months=l)
-        )
-        # leadtime period end
-        fdate = (
-            date.fromisoformat(f'{btime.year}-'+
-                                f'{btime.month:02d}-'+
-                                f'{1:02d}')
-            +relativedelta(months=l+period_length)
-        )
-        # store array of forecast 3 month period
-        forecasttime.append(f'{sdate.strftime("%b")}-{fdate.strftime("%b %Y")}')
-
-    # construct forecast period only during the binned period
-    mean_forecasttime = [forecasttime[idx] for idx in bins[:-1]]
-
-    # get the initial time
-    ini_time_date = (
-        date.fromisoformat(
-            f'{btime.year}-'+
-            f'{btime.month:02d}-'+
-            f'{1:02d}'
-        )
-    )
-    # construct the initial time format
-    ini_time = f'{ini_time_date.strftime("%b %Y")}'
-
-    return {'init':ini_time,'fcsts':mean_forecasttime}
 
 if __name__=='__main__':
 
@@ -245,13 +27,14 @@ if __name__=='__main__':
     data_source = 'regrid'
 
     # getting the regionl mask on mom grid
-    ds_lme = get_mom6_regionl_mask()
+    ds_lme = mp.MOM6Static.get_mom6_regionl_mask()
 
     # getting the ocean mask on mom grid
-    da_lmask = get_mom6_mask(mask='wet',source=data_source)
+    da_lmask = mp.MOM6Static.get_mom6_mask(mask='wet',source=data_source)
 
     # loaded the mom6 raw field
-    ds_data = get_mom6(iyear=ini_year,imonth=ini_month,var=varname,source=data_source)
+    mom6Forecast = mp.MOM6Forecast(iyear=ini_year,imonth=ini_month,var=varname,source=data_source)
+    ds_data = mom6Forecast.get_mom6()
 
     # load variable to memory
     da_data = ds_data[varname].isel(init=0)
@@ -281,7 +64,7 @@ if __name__=='__main__':
 
     # load the predetermined hindcast/forecast tercile value
     #  this is based on 30 years statistic 1993-2023
-    ds_tercile = get_mom6_tercile(imonth=ini_month,var=varname,source=data_source)
+    ds_tercile = mom6Forecast.get_mom6_tercile()
 
     # average the forecast over the lead bins
     ds_tercile_binned = (
@@ -332,11 +115,7 @@ if __name__=='__main__':
 
     ############# plotting the max tercile probability
     # get time format
-    dict_forecast_time = get_init_fcst_time(
-        iyear=ini_year,
-        imonth=ini_month,
-        bins=lead_bins
-    )
+    dict_forecast_time = mom6Forecast.get_init_fcst_time()
 
     # create the new colormap designed for tercile plot
     ncolor = 18
@@ -425,4 +204,11 @@ if __name__=='__main__':
         fontsize=16
     )
 
-    fig.savefig('/home/chsu/regional_mom/figures/tercile_regrid.png')
+    proj_dir = os.path.dirname(
+        os.path.dirname(
+        os.path.dirname(
+            os.path.realpath(__file__)
+        )
+        )
+    )
+    fig.savefig(f'{proj_dir}/figures/tercile_regrid.png')
